fix(md-tags): 🐛 change .cm-hashtag from inline-block to inline.
For tags, the hash character is in a separate span element, which, together with the display: inline-block property can lead to an incorrect line break between the hash character and the tag text. If we use display: inline instead of inline-block, the normal text-breaking rules apply, and all is well.

closes  ceciliamay/obsidianmd-theme-primary#73